Started rationalizing Exchange API and impl class locations
